Supporting maongo-java-server rather than fongo
