+++ pom.xml
@@ -6,7 +6,7 @@
 
 	<groupId>com.aidanwhiteley</groupId>
 	<artifactId>books</artifactId>
-	<version>0.4.1-SNAPSHOT</version>
+	<version>0.4.2-SNAPSHOT</version>
 	<packaging>jar</packaging>
 
 	<name>Books Microservice</name>
