+++ src/main/resources/application-mongo-java-server.yml
@@ -1,0 +1,33 @@
+spring:
+  data:
+    mongodb:
+      uri: mongodb://localhost:27017/books-mongo-java-server?safe=true&journal=true&w=majority&connectTimeoutMS=1001&wtimeoutMS=1000&serverSelectionTimeoutMS=10000
+  mail:
+    port: 3025    # Must match the value specified in the SignUpNotificationServiceTest. Set hight to avoid problems on Unix envs
+
+books:
+  client:
+    enableCORS: true      # Note - this value overriden in the IntegrationTest base test class
+    allowedCorsOrigin: http://localhost:9000
+    postLogonUrl: http://localhost:9000/
+    xsrfHeader: X-XSRF
+
+  reload:
+    development:
+      data: true    # Whether the contents of the /src/main/resources/sample_data will be reloaded into the database at application startup. Deletes all existing data!!!
+
+  users:
+    default:
+      admin:
+        email: example@exmaple.com  # The remote email address for the user that will be automatically made an admin on first logon
+    registrationAdminEmail:
+      emailFrom: exampleFrom@example.com
+      emailTo: exampleTo@example.com
+
+  jwt:
+    secretKey: This^IsNot!A%Real*Secret
+    cookieOverHttpsOnly: false
+    cookieAccessedByHttpOnly: true
+
+  oauth2:
+    cookieOverHttpsOnly: false
