+++ src/main/resources/banner.txt
@@ -1,9 +1,10 @@
 
-    ____              __           __  ____                                      _
-   / __ )____  ____  / /_______   /  |/  (_)_____________  ________  ______   __(_)_______
-  / __  / __ \/ __ \/ //_/ ___/  / /|_/ / / ___/ ___/ __ \/ ___/ _ \/ ___/ | / / / ___/ _ \
- / /_/ / /_/ / /_/ / ,< (__  )  / /  / / / /__/ /  / /_/ (__  )  __/ /   | |/ / / /__/  __/
-/_____/\____/\____/_/|_/____/  /_/  /_/_/\___/_/   \____/____/\___/_/    |___/_/\___/\___/
+______             _         ___  ____                                    _
+| ___ \           | |        |  \/  (_)                                  (_)
+| |_/ / ___   ___ | | _____  | .  . |_  ___ _ __ ___  ___  ___ _ ____   ___  ___ ___
+| ___ \/ _ \ / _ \| |/ / __| | |\/| | |/ __| '__/ _ \/ __|/ _ \ '__\ \ / / |/ __/ _ \
+| |_/ / (_) | (_) |   <\__ \ | |  | | | (__| | | (_) \__ \  __/ |   \ V /| | (_|  __/
+\____/ \___/ \___/|_|\_\___/ \_|  |_/_|\___|_|  \___/|___/\___|_|    \_/ |_|\___\___|
 
 
  Aidan Whiteley - https://github.com/aidanwhiteley/books
