+++ src/test/resources/logback.xml
@@ -9,6 +9,7 @@
     <logger name="org.mongodb" level="WARN"/>
     <logger name="com.icegreen.greenmail" level="WARN" />
     <logger name="springfox.documentation" level="WARN"/>
+    <logger name="de.bwaldvogel.mongo" level="WARN"/>
 
     <logger name="com.aidanwhiteley" level="WARN"/>
     <logger name="com.aidanwhiteley.books.controller" level="WARN" />
